+++ modules/http/CMakeLists.txt
@@ -17,6 +17,8 @@ endif ()
 set(HTTP_DESTINATION_SOURCES
     http.h
     http.c
+    http-worker.h
+    http-worker.c
     http-parser.c
     http-parser.h
     http-plugin.c
