+++ modules/http/http.h
@@ -30,17 +30,6 @@
 
 #include "logthrdestdrv.h"
 
-#define CURL_NO_OLDIES 1
-#include <curl/curl.h>
-
-typedef struct _HTTPDestinationWorker
-{
-  LogThreadedDestWorker super;
-  CURL *curl;
-  GString *request_body;
-  struct curl_slist *request_headers;
-} HTTPDestinationWorker;
-
 typedef struct
 {
   LogThreadedDestDriver super;
